Fixed 42 norm errors

+++ philo/src/time.c
@@ -1,3 +1,15 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   time.c                                             :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: vfries <vfries@student.42lyon.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/03/16 07:14:46 by vfries            #+#    #+#             */
+/*   Updated: 2023/03/16 07:14:49 by vfries           ###   ########lyon.fr   */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include <sys/time.h>
 
 #include "philosophers.h"
@@ -20,13 +32,13 @@ struct timeval	get_current_time(void)
 long long	get_timestamp(const t_philosopher *philosopher,
 				const struct timeval current_time)
 {
-	return((long long)(current_time.tv_sec - philosopher->start_time.tv_sec)
-		* NB_OF_USECONDS_IN_A_MILLISECOND
+	return ((current_time.tv_sec - philosopher->start_time.tv_sec)
+		* (long long)NB_OF_USECONDS_IN_A_MILLISECOND
 		+ (current_time.tv_usec - philosopher->start_time.tv_usec)
 		/ NB_OF_USECONDS_IN_A_MILLISECOND);
 }
 
-void timeval_add_ms(struct timeval *tv, const int number_of_ms_to_add)
+void	timeval_add_ms(struct timeval *tv, const int number_of_ms_to_add)
 {
 	tv->tv_usec += number_of_ms_to_add * NB_OF_USECONDS_IN_A_MILLISECOND;
 	if (tv->tv_usec >= NB_OF_USECONDS_IN_A_SECOND)
@@ -36,7 +48,7 @@ void timeval_add_ms(struct timeval *tv, const int number_of_ms_to_add)
 	}
 }
 
-int timeval_compare(const struct timeval t1, const struct timeval t2)
+int	timeval_compare(const struct timeval t1, const struct timeval t2)
 {
 	if (t1.tv_sec < t2.tv_sec)
 		return (-1);
